requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
defusedxml==0.7.1
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-anthropic==0.3.0
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langgraph-checkpoint==2.0.1
langgraph-sdk==0.1.33
langsmith==0.1.147
moviepy==1.0.3
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
openai==1.52.0
orjson==3.10.9
packaging==24.1
pandas==2.2.3
Pillow==9.5.0
proglog==0.1.10
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
yt-dlp==2024.12.3
